feat: modal for ticket reason
